Move mapif functions from int_quest.c to mapif.c
